Cleaning and moving props part 16
